ml-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic==2.7.4
python-socketio==5.11.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
scipy==1.13.1
surprise==1.1.3
httpx==0.27.0
asyncpg==0.29.0
redis==5.0.8
python-dotenv==1.0.1
